packepfecam/backend/requirements.txt
anyio
autopep8
bcrypt
beanie
certifi
cffi
click
cryptography
dnspython
email-validator
fastapi
h11
httpcore
httptools
httpx
idna
itsdangerous
Jinja2
lazy-model
MarkupSafe
motor
orjson
passlib
pycodestyle
pycparser
pydantic
PyJWT
pymongo
python-dotenv
python-multipart
PyYAML
rfc3986
six
sniffio
starlette
toml
tomli
typing_extensions
ujson
uvicorn
uvloop
watchfiles
websockets
opencv-python
ultralytics
PyOpenGL
aiofiles
mongoengine
pytest
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]